requirements.txt
arxiv==1.4.7
requests==2.31.0
python-dotenv==1.0.0
huggingface-hub==0.19.4
aiohttp==3.9.1
pandas==2.1.3
SQLAlchemy==2.0.23
python-dateutil==2.8.2